requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pypdf
python-multipart
pydantic
tqdm
streamlit
requests
python-dotenv
langchain-core
langchain-google-genai
google-generativeai
